riscv: Add basic extension support for XTheadFmv and XTheadInt
This patch add basic support for the XTheadFmv and XTheadInt
ISA extension. As both extensions only contain instruction,
which are not supposed to be emitted by the compiler, the support
only covers awareness of the extension name in the march string
and the definition of a feature test macro.

gcc/ChangeLog:

	* common/config/riscv/riscv-common.cc: Add xtheadfmv and
	  xtheadint.
	* config/riscv/riscv-opts.h (MASK_XTHEADMAC): New.
	(MASK_XTHEADFMV): New.
	(TARGET_XTHEADFMV): New.
	(MASK_XTHEADINT): New.
	(TARGET_XTHEADINT): New.
	(MASK_XTHEADMEMIDX): New.
	(MASK_XTHEADSYNC): New.

gcc/testsuite/ChangeLog:

	* gcc.target/riscv/xtheadfmv.c: New test.
	* gcc.target/riscv/xtheadint.c: New test.

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>